Final reorganization and polish
